requirements.txt
Authlib==1.4.1
boto3==1.36.23
bs4==0.0.2
email-validator==2.2.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
flask-jwt-extended==4.7.1
Flask-Limiter==3.10.1
flask-login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restful==0.3.10
Flask-SeaSurf==2.0.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
gunicorn==23.0.0
hvac==2.3.0
httpx==0.28.1
matplotlib==3.10.0
# openai-whisper==20240930  # Comment out or remove this line if not essential
#openai-whisper
openai
pdf2image
pdfkit
pdfplumber
pillow
phonenumbers
praw
psycopg2
pycryptodome
pydantic
PyMuPDF
pymongo[srv]
pymysql
PyPDF2[full]
pyotp
pytesseract
python-docx
python-dotenv
pytube
requests
unidecode
WTForms
yfinance
kagglehub[pandas-datasets]

